Fixed Barrel
The barrel was using ints instead of floats.

+++ Assets/Scripts/Barrel.cs
@@ -6,16 +6,16 @@ public class Barrel : MonoBehaviour, IDamage
 {
     [SerializeField] ExplosionImplosion explosion;
     [SerializeField] MeshCollider collider;
-    [SerializeField] int health;
+    [SerializeField] float health;
 
-    int maxhealth;
+    float maxhealth;
     // Start is called before the first frame update
     void Start()
     {
         maxhealth = health;
     }
 
-    public void takeDamage(int dmg)
+    public void takeDamage(float dmg)
     {
         health -= dmg;
